add 'machine' package for controllers and machine providers
