--- app/src/main/java/com/pileproject/drive/programming/visual/block/selection/IfMachineWasTouchedBlock.java
@@ -1,37 +1,0 @@
-/**
- * Copyright (C) 2011-2016 PILE Project, Inc. <dev@pileproject.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.pileproject.drive.programming.visual.block.selection;
-
-import android.content.Context;
-
-import com.pileproject.drive.R;
-import com.pileproject.drive.execution.CarControllerBase;
-import com.pileproject.drive.execution.MachineController;
-
-/**
- * This block checks the touch sensor was touched or not
- */
-public class IfMachineWasTouchedBlock extends SelectionBlock {
-
-    public IfMachineWasTouchedBlock(Context context) {
-        super(context, R.layout.block_if_machine_was_touched);
-    }
-
-    @Override
-    protected boolean evaluateCondition(MachineController controller) {
-        return ((CarControllerBase) controller).isTouchSensorTouched();
-    }
-}
