+++ app/src/main/java/com/pileproject/drive/programming/visual/block/selection/car/IfThereWasALargeSoundBlock.java
@@ -13,35 +13,36 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.pileproject.drive.programming.visual.block.selection;
+package com.pileproject.drive.programming.visual.block.selection.car;
 
 import android.content.Context;
 
 import com.pileproject.drive.R;
-import com.pileproject.drive.execution.CarControllerBase;
-import com.pileproject.drive.execution.MachineController;
+import com.pileproject.drive.machine.CarControllerBase;
+import com.pileproject.drive.machine.MachineController;
 import com.pileproject.drive.preferences.BlockPreferences;
+import com.pileproject.drive.programming.visual.block.selection.SelectionBlock;
+
 
 /**
- * This block check the light sensor's value
- * If the value is higher than threshold, it means the machine is on the
- * light-colored floor and out of black line.
+ * This block check the sound sensor's value and check there
+ * was a large sound or not
  *
  * @author <a href="mailto:tatsuyaw0c@gmail.com">Tatsuya Iwanari</a>
  * @version 1.0 7-July-2013
  */
-public class IfMachineIsOutOfLineBlock extends SelectionBlock {
+public class IfThereWasALargeSoundBlock extends SelectionBlock {
 
     private int mThreshold;
 
-    public IfMachineIsOutOfLineBlock(Context context) {
-        super(context, R.layout.block_if_machine_is_out_of_line);
+    public IfThereWasALargeSoundBlock(Context context) {
+        super(context, R.layout.block_if_there_was_a_large_sound);
 
-        mThreshold = BlockPreferences.get(context).getLineSensorThreshold();
+        mThreshold = BlockPreferences.get(context).getSoundSensorThreshold();
     }
 
     @Override
     protected boolean evaluateCondition(MachineController controller) {
-        return ((CarControllerBase) controller).getLineSensorValue() > mThreshold;
+        return ((CarControllerBase) controller).getSoundSensorDb() > mThreshold;
     }
 }
