+++ app/src/main/java/com/pileproject/drive/programming/visual/block/sequence/car/SetRightMotorSpeedBlock.java
@@ -13,35 +13,38 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.pileproject.drive.programming.visual.block.sequence;
+package com.pileproject.drive.programming.visual.block.sequence.car;
 
 import android.content.Context;
 
 import com.pileproject.drive.R;
-import com.pileproject.drive.execution.CarControllerBase;
+import com.pileproject.drive.machine.CarControllerBase;
 import com.pileproject.drive.execution.ExecutionCondition;
-import com.pileproject.drive.execution.MachineController;
+import com.pileproject.drive.machine.MachineController;
+import com.pileproject.drive.programming.visual.block.sequence.SequenceBlockHasNumberText;
 import com.pileproject.drive.util.development.Unit;
 import com.pileproject.drive.util.math.Range;
 
 import java.math.BigDecimal;
 
+import static com.pileproject.drive.machine.CarControllerBase.MotorKind.RightMotor;
+
 /**
- * Forward for a while
+ * Set right motor power
  *
  * @author yusaku
  * @version 1.0 7-July-2013
  */
-public class ForwardSecBlock extends SequenceBlockHasNumberText {
+public class SetRightMotorSpeedBlock extends SequenceBlockHasNumberText {
 
     // TODO: set from preference
-    private static final Range<BigDecimal> range = Range.closed(BigDecimal.ZERO, new BigDecimal(3));
+    private static final Range<BigDecimal> range = Range.closed(BigDecimal.ZERO, new BigDecimal(100));
 
     // TODO: set from preference
-    private static final int PRECISION = 3;
+    private static final int PRECISION = 0;
 
-    public ForwardSecBlock(Context context) {
-        super(context, R.layout.block_forward_sec, R.id.block_numText);
+    public SetRightMotorSpeedBlock(Context context) {
+        super(context, R.layout.block_set_right_motor_speed, R.id.block_numText);
     }
 
     @Override
@@ -56,12 +59,12 @@ public class ForwardSecBlock extends SequenceBlockHasNumberText {
 
     @Override
     public int action(MachineController controller, ExecutionCondition condition) {
-        ((CarControllerBase) controller).moveForward();
-        return getActionValue();
+        ((CarControllerBase) controller).setMotorPower(RightMotor, getValue().intValue());
+        return 0;
     }
 
     @Override
     public Unit getUnit() {
-        return Unit.Second;
+        return Unit.Percentage;
     }
 }
