+++ app/src/nxt/java/com/pileproject/drive/machine/NxtMachineProvider.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.pileproject.drive.execution;
+package com.pileproject.drive.machine;
 
 import com.pileproject.drivecommand.machine.MachineBase;
 import com.pileproject.drivecommand.model.com.ICommunicator;
