+++ app/src/nxt/java/com/pileproject/drive/setting/machine/NxtThresholdFragment.java
@@ -29,8 +29,8 @@ import android.widget.SeekBar.OnSeekBarChangeListener;
 import android.widget.TextView;
 
 import com.pileproject.drive.R;
-import com.pileproject.drive.execution.NxtCarController.SensorProperty.LineSensor;
-import com.pileproject.drive.execution.NxtCarController.SensorProperty.SoundSensor;
+import com.pileproject.drive.machine.NxtCarController.SensorProperty.LineSensor;
+import com.pileproject.drive.machine.NxtCarController.SensorProperty.SoundSensor;
 import com.pileproject.drive.preferences.BlockPreferences;
 
 public class NxtThresholdFragment extends DialogFragment {
